+++ app/src/main/java/com/example/model/Message.java
@@ -1,32 +1,41 @@
 package com.example.model;
 
+
 public class Message {
-    private int id;
     private int conversation_id;
     private int sender_id;
     private String message;
-    private String created_at;
+    private long created_at;
 
-    public Message(int conversation_id, int sender_id, String message) {
+    public Message(){
+
+    }
+
+    public Message(Integer user_id, Integer conversation_id, String body) {
+        sender_id = user_id;
+        message = body;
+        created_at = 0;
         this.conversation_id = conversation_id;
-        this.sender_id = sender_id;
-        this.message = message;
     }
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
+    public String getMessage() {
+        return message;
     }
 
     public int getConversation_id() {
         return conversation_id;
     }
 
-    public void setConversation_id(int conversation_id) {
-        this.conversation_id = conversation_id;
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public long getCreated_at() {
+        return created_at;
+    }
+
+    public void setCreated_at(long created_at) {
+        this.created_at = created_at;
     }
 
     public int getSender_id() {
@@ -37,19 +46,7 @@ public class Message {
         this.sender_id = sender_id;
     }
 
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public String getCreated_at() {
-        return created_at;
-    }
-
-    public void setCreated_at(String created_at) {
-        this.created_at = created_at;
+    public void setConversation_id(int conversation_id) {
+        this.conversation_id = conversation_id;
     }
 }
