+++ app/src/main/java/com/example/tinder/message_box/MatchListAdapter.java
@@ -7,10 +7,6 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.example.tinder.R;
 
@@ -25,8 +21,6 @@ public class MatchListAdapter extends RecyclerView.Adapter<MatchListAdapter.Matc
     }
 
     public static class  MatchViewHolder extends RecyclerView.ViewHolder {
-        public TextView textView;
-        LinearLayout line;
         public MatchViewHolder(View view) {
             super(view);
         }
@@ -42,8 +36,7 @@ public class MatchListAdapter extends RecyclerView.Adapter<MatchListAdapter.Matc
 
     @Override
     public void onBindViewHolder(@NonNull MatchViewHolder matchViewHolder, int i) {
-//        TextView textView = matchViewHolder.textView;
-//        textView.setText(this.datas.get(i));
+
     }
 
     @Override
